server/requirements.txt
alipay-sdk-python==3.1.6
astroid==2.0.3
Django==2.1
image==1.5.24
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.13
Pillow==5.2.0
pylint==2.1.1
pylint-django==2.0
pylint-plugin-utils==0.4
pytz==2018.5
qcloudsms-py==0.1.3
qrcode==6.0
six==1.11.0
wrapt==1.10.11
coverage==4.5.1
openpyxl==2.5.5
xlrd==1.1.0
xlwt==1.3.0
Sphinx==1.7.6
Sphinxcontrib-websupport==1.1.0
Sphinx-rtd-theme==0.4.1